Remove most pre-generated bindings and generate them at compile time instead
This introduces the require_gobject(typelib_name) macro, which replaces
itself with the full generated library for the given typelib name.
This will allow to
- create shards that provide hand-written wrappers around GIR libraries
  while still making use of this for the bulk of the work
- split this into the pure GIR parts and shards that provide the
  convenience and patchwork (GLib and GObject patchwork will likely stay here)
- End users to use GIR libraries that this does not ship pregenerated
  wrappers for in a more convenient manner

It still requires the GIRepository bindings to work, which require
the GLib and GObject lib bindings, so we keep those pregenerated
and the old generator to update them.

The shims now use the macro instead of requiring the pregenerated things.

Also adds a symlink from lib/gobject to src/ so that the samples
can use the same requires that a user of this shard would.