More idiomatic wrappers for methods taking a collection argument
In Crystal we typically know the size of our collections.
But even more so we already have to allocate a new array for
unwrapping the collection items before passsing anyways. So we can
just take the size of that array and pass it to methods that need the size
of their array parameter passed.

So this now hides the size parameters from the wrapper definitions and computes
the size internally in the wrapper. The heuristik for this is for now is just
looking for n_foo for a given foo argument of type array